@@ -1,6 +1,6 @@
 {
   "name": "jupyter_firefly_extensions",
-  "version": "0.2.6",
+  "version": "2.0.0",
   "description": "A package for rendering FITS in Jupyter",
   "author": "Trey Roby <roby@ipac.caltech.edu>",
   "main": "lib/lab_extension.js",
@@ -20,40 +20,45 @@
     "watch:lib": "babel src -d lib --watch",
     "watch": "npm-run-all -p watch:*"
   },
-  "babel": {
-    "presets": [
-      "latest"
-    ],
-    "plugins": [
-      "transform-class-properties"
-    ]
-  },
   "jupyterlab": {
     "extension": true
   },
   "dependencies": {
     "b64-to-blob": "^1.2.19",
+    "regenerator-runtime": "~0.13",
     "lodash": "^4.17",
     "firefly-api-access": "0.0.4",
-    "@jupyterlab/docregistry": "^0.18.0"
+    "@jupyterlab/docregistry": "^2.2"
   },
   "devDependencies": {
-    "babel-cli": "^6.26.0",
-    "babel-core": "^6.18.2",
-    "babel-loader": "^6.4.0",
+    "@babel/cli": "^7.10",
+    "@babel/core": "^7.8",
+    "@babel/preset-env": "~7.8",
     "babel-plugin-transform-class-properties": "^6.19.0",
-    "babel-preset-latest": "^6.16.0",
-    "css-loader": "^0.25.0",
+    "@babel/plugin-transform-runtime": "~7.8",
+    "css-loader": "^3.4",
     "file-loader": "^0.9.0",
     "json-loader": "^0.5.4",
     "mkdirp": "^0.5.1",
     "npm-run-all": "^4.1.1",
     "rimraf": "^2.6.2",
-    "style-loader": "^0.13.1",
-    "url-loader": "^1.1.1",
-    "webpack": "^2.2.0"
+    "style-loader": "^1.1",
+    "url-loader": "^2.3",
+    "webpack": "^4.41",
+    "eslint": "~6.8",
+    "babel-eslint": "~10.0",
+    "eslint-loader": "~3.0",
+    "eslint-plugin-jsdoc": "~21.0",
+    "eslint-plugin-jsx-a11y": "~6.2",
+    "eslint-plugin-react": "~7.18",
+    "eslint-plugin-react-hooks": "~3.0"
+  },
+  "optionalDependencies": {
+    "fsevents": "~2.1"
   },
   "resolutions": {
-    "**/event-stream": "^4.0.1"
+    "**/event-stream": "^4.0.1",
+    "**/chokidar/fsevents": "~2.1",
+    "**/jest-haste-map/fsevents": "~2.1"
   }
 }
