@@ -5,6 +5,6 @@
 # Distributed under the terms of the Modified BSD License.
 
 
-version_info = (0, 2, 6)
+version_info = (2, 0, 0)
 __version__ = ".".join(map(str, version_info))
 version_short = ".".join(map(str,version_info[0:3]))
